refactor(lowering): extract function frame setup
