+++ scripts/source-file-size-allowlist.json
@@ -1,5 +1,3 @@
 {
-  "hardCap": {
-    "src/lowering/functionLowering.ts": 1014
-  }
+  "hardCap": {}
 }
